requirements.txt
torch
diffusers
supabase
Pillow
accelerate
clean-fid
scipy
numpy
opencv-python
controlnet_aux
requests